+++ src/main/scala/com/yogpc/qp/machines/marker/TileFlexMarker.java
@@ -60,16 +60,16 @@ public class TileFlexMarker extends BlockEntity implements QuarryMarker, Checker
             int d = getDistance(max, offset, facing.getAxis());
             if (d > maxRange) {
                 max = getLimited(max, offset, facing, maxRange);
-            } else if (d < 1) {
-                max = getLimited(max, offset, facing, 1);
+            } else if (d < 0) {
+                max = getLimited(max, offset, facing, 0);
             }
         } else {
             min = min.relative(facing, amount);
             int d = getDistance(offset, min, facing.getAxis());
             if (d > maxRange) {
                 min = getLimited(min, offset, facing, maxRange);
-            } else if (d < 1) {
-                min = getLimited(min, offset, facing, 1);
+            } else if (d < 0) {
+                min = getLimited(min, offset, facing, 0);
             }
             if (facing == Direction.DOWN && min.getY() < level.getMinBuildHeight()) {
                 min = new BlockPos(min.getX(), level.getMinBuildHeight(), min.getZ());
@@ -128,7 +128,11 @@ public class TileFlexMarker extends BlockEntity implements QuarryMarker, Checker
 
     @Override
     public Optional<Area> getArea() {
-        return Optional.of(new Area(min, max, direction));
+        if (this.direction == null) {
+            return Optional.empty();
+        } else {
+            return Optional.of(new Area(min, max, direction));
+        }
     }
 
     @Override
@@ -171,6 +175,12 @@ public class TileFlexMarker extends BlockEntity implements QuarryMarker, Checker
 
         public static Movable valueOf(int i) {
             return v[i];
+        }
+
+        public int distanceFromOrigin(BlockPos origin, BlockPos areaMin, BlockPos areaMax, Direction facing) {
+            Direction actualFacing = getActualFacing(facing);
+            BlockPos relative = actualFacing.getAxisDirection() == Direction.AxisDirection.POSITIVE ? areaMax : areaMin;
+            return Math.abs(getDistance(origin, relative, actualFacing.getAxis()));
         }
 
         static {